+++ src/SelectField/index.js
@@ -5,7 +5,7 @@ import css from 'styled-jsx/css'
 import { ArrowDown } from '../icons/Arrow'
 import { LabelFilled, LabelOutlined } from '../FieldLabel'
 import { Select } from './Select'
-import { colors } from '../theme'
+import { colors } from '../theme.js'
 
 const TailIcon = () => (
     <div>
@@ -42,7 +42,7 @@ class SelectField extends React.Component {
     }
 
     isFocused() {
-        return this.state.focused
+        return this.state.focus
     }
 
     render() {
@@ -60,9 +60,6 @@ class SelectField extends React.Component {
             loading,
             children,
             value,
-            autocomplete,
-            autofocus,
-            readonly,
             tabIndex,
             onChange,
         } = this.props
@@ -85,15 +82,13 @@ class SelectField extends React.Component {
                 dense={dense}
             >
                 <Select
+                    focus={focus}
                     name={name}
                     value={value}
                     disabled={disabled}
                     filled={filled}
                     dense={dense}
-                    autocomplete={autocomplete}
-                    autofocus={autofocus}
                     tabIndex={tabIndex}
-                    readonly={readonly}
                     onChange={onChange}
                     onFocus={this.onFocus}
                     onBlur={this.onBlur}
@@ -114,7 +109,6 @@ SelectField.propTypes = {
 
     value: propTypes.string,
     className: propTypes.string,
-    autocomplete: propTypes.string,
     tabIndex: propTypes.string,
 
     children: propTypes.oneOfType([
@@ -136,8 +130,6 @@ SelectField.propTypes = {
     warning: propTypes.bool,
     error: propTypes.bool,
     loading: propTypes.bool,
-    autofocus: propTypes.bool,
-    readonly: propTypes.bool,
     initialFocus: propTypes.bool,
 
     onFocus: propTypes.func,