+++ src/index.js
@@ -1,41 +1,33 @@
+/* global */
+export { CssReset } from './CssReset'
+
+/* atoms */
 export { AlertBar } from './AlertBar'
 export { AlertStack } from './AlertStack'
 export { Button } from './Button'
-export { ButtonStrip } from './ButtonStrip'
 export { Card } from './Card'
 export { Checkbox } from './Checkbox'
 export { Chip } from './Chip'
-export { CircularProgress } from './CircularProgress'
+export { CircularLoader } from './CircularLoader'
+export { ComponentCover } from './ComponentCover'
 export { Divider } from './Divider'
-export { DropdownButton } from './DropdownButton'
-
 export { Help } from './Help'
 export { InputField } from './InputField'
-export { LinearProgress } from './LinearProgress'
-
+export { LinearLoader } from './LinearLoader'
 export { Logo, LogoIcon, LogoIconWhite, LogoWhite } from './Logo'
-
-export { Menu } from './Menu'
 export { MenuItem } from './MenuItem'
-export { Modal } from './Modal'
-
 export { Radio } from './Radio'
-export { SelectField } from './SelectField'
-export { SplitButton } from './SplitButton'
-export { Switch } from './Switch'
-export { Node } from './Node'
-
-export { CssReset } from './CssReset'
-
-export { theme, colors } from './theme.js'
-
-export { ComponentCover } from './ComponentCover'
 export { ScreenCover } from './ScreenCover'
+export { SelectField } from './SelectField'
+export { Switch } from './Switch'
 
-/*
- * Components below are internal and may be exposed at some point, but
- * are kept internal for now until the API stabilizes
- */
+/* molecules */
+export { ButtonStrip } from './ButtonStrip'
+export { DropdownButton } from './DropdownButton'
+export { Menu } from './Menu'
+export { Node } from './Node'
+export { Modal } from './Modal'
+export { SplitButton } from './SplitButton'
 
-// export { DropMenu } from './DropMenu'
-// export { LabelFilled, LabelOutlined } from './FieldLabel'
+/* constants */
+export { theme, colors } from './theme.js'